enrico's script push

+++ my-auto-setup.sh
@@ -1,23 +1,24 @@
 #!/usr/bin/env bash
 ##!/bin/sh
 
-echo "\033[0;36m Welcome! 🦆 \n my-auto-setup started! 🚀 \033[0m"
+echo "\033[0;36m Welcome! 🦆 \n my-automated-mac-setup started! 🚀 \033[0m"
 
-# TODO trasformare tutti i commenti in echo in inglese :)
 # TODO trascrivere a commento le descrizioni dello stato del mac dagli screen :)
+
+echo "\033[0;34m Setting to 0 the wait time for showing the dock ⏲ \033[0m"
 defaults write com.apple.dock autohide-delay -float 0; defaults write com.apple.dock autohide-time-modifier -int 0;killall Dock
 # undo: defaults write com.apple.dock autohide-delay -float 0.5; defaults write com.apple.dock autohide-time-modifier -int 0.5 ;killall Dock
 
-# 𝗗𝗶𝘀𝗮𝗯𝗹𝗲 𝗔𝗻𝗻𝗼𝘆𝗶𝗻𝗴 𝗗𝗶𝘀𝗸 𝗪𝗮𝗿𝗻𝗶𝗻𝗴 (must restart Mac to take effect): 
+echo "\033[0;34m Disabling annoying disk warning when unmounting external devices 💾 \033[0m"
 sudo defaults write /Library/Preferences/SystemConfiguration/com.apple.DiskArbitration.diskarbitrationd.plist DADisableEjectNotification -bool YES && sudo pkill diskarbitrationd
 echo -e "\033[1;31m Must restart Mac to take effect 🔁 \033[0m"
 # 𝗥𝗲-𝗘𝗻𝗮𝗯𝗹𝗲 𝗔𝗻𝗻𝗼𝘆𝗶𝗻𝗴 𝗗𝗶𝘀𝗸 𝗪𝗮𝗿𝗻𝗶𝗻𝗴:
 # sudo defaults delete /Library/Preferences/SystemConfiguration/com.apple.DiskArbitration.diskarbitrationd.plist DADisableEjectNotification && sudo pkill diskarbitrationd
 
-# 𝗖𝗵𝗮𝗻𝗴𝗲 𝗦𝗰𝗿𝗲𝗲𝗻𝘀𝗵𝗼𝘁 𝗗𝗲𝗳𝗮𝘂𝗹𝘁 𝘁𝗼 𝗝𝗣𝗚 (replace with png to undo):
+echo "\033[0;34m Changing screenshot default to jpg (replace with png to undo) 📸 \033[0m"
 defaults write com.apple.screencapture type jpg
 
-# 𝗠𝗮𝗸𝗲 𝗛𝗶𝗱𝗱𝗲𝗻 𝗔𝗽𝗽𝘀 𝗧𝗿𝗮𝗻𝘀𝗽𝗮𝗿𝗲𝗻𝘁:
+echo "\033[0;34m Making hidden apps transparent 🫥 \033[0m"
 defaults write com.apple.Dock showhidden -bool TRUE && killall Dock
 
 
@@ -49,7 +50,7 @@ defaults write com.apple.Dock showhidden -bool TRUE && killall Dock
 
 # TODO trascrivi come commenti la disposizione delle app nel launchpad
 
-# some apps without brew script I have to manually download :(
+echo "\033[0;34m Opening some link to apps you have to manually download (no brew script) 🛠 \033[0m"
 open https://apps.apple.com/us/app/accelerate-for-safari/id1459809092 
 open https://apps.apple.com/it/app/piper/id1421915518?mt=12
 open https://apps.apple.com/it/app/enki-learn-coding-programming/id993753145
@@ -77,102 +78,50 @@ sudo open -a /Applications/Install\ Visual\ Studio\ for\ Mac.app
 hdiutil unmount /Volumes/Visual\ Studio\ for\ Mac\ Installer
 rm Downloads/visualstudioformacinstaller-*.dmg
 
-# installing homebrew :)
+echo "\033[0;34m Installing Homebrew! 🍺 \033[0m"
 /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
 
-# and all the other apps i need (also opening them to setup some preferences)
-brew install --cask onedrive
-sudo open -a onedrive
-
-brew install --cask aldente
-sudo open -a aldente
-
+echo "\033[0;34m Installing all the other apps i need (also I'll open them to setup some preferences) 💻🖥 \033[0m"
+brew install --cask onedrive; sudo open -a onedrive
+brew install --cask aldente; sudo open -a aldente
 brew install git
 git config --global user.email "enrico.pezzano@outlook.it"
 git config --global user.name "EnricoPezzano"
-
-brew install --cask microsoft-teams
-sudo open -a "microsoft teams"
-
-brew install --cask whatsapp
-sudo open -a whatsapp
-
-brew install --cask telegram
-sudo open -a telegram
-
-brew install --cask adguard-vpn
-sudo open -a "adguard vpn"
-
-brew install --cask iterm2
-sudo open -a iterm2
-
-brew install --cask rectangle
-sudo open -a rectangle
-
+brew install --cask microsoft-teams; sudo open -a "microsoft teams"
+brew install --cask whatsapp; sudo open -a whatsapp
+brew install --cask telegram; sudo open -a telegram
+brew install --cask adguard-vpn; sudo open -a "adguard vpn"
+brew install --cask iterm2; sudo open -a iterm2
+brew install --cask rectangle; sudo open -a rectangle
 brew install python #-tk@3.9
 brew install python-tk@3.9
-
-brew install --cask google-chrome
-sudo open -a "Google Chrome" 
+brew install --cask google-chrome; sudo open -a "Google Chrome" 
 brew install --cask chromedriver
-
 brew install --cask mamp
 rm -rf ../../Applications/MAMP\ PRO.app
 sudo open -a mamp
-
-brew install --cask utm
-sudo open -a utm
-
+brew install --cask utm; sudo open -a utm
 brew install --cask vlc
-
-brew install --cask firefox
-sudo open -a firefox
-
-brew install --cask visual-studio-code
-sudo open -a "visual studio code"
-
-brew install --cask 4k-video-downloader
-
-brew install --cask discord
-sudo open -a discord
-
-brew install --cask intellij-idea
-sudo open -a "intellij idea"
-
-brew install --cask alt-tab
-sudo open -a alttab
-
-brew install --cask cheatsheet
-sudo open -a cheatsheet
-
-brew install --cask appcleaner
-sudo open -a appcleaner
-
-brew install --cask lunar
-sudo open -a lunar
-
+brew install --cask firefox; sudo open -a firefox
+brew install --cask visual-studio-code; sudo open -a "visual studio code"
+brew install --cask 4k-video-downloader; sudo open -a "4k video downloader"
+brew install --cask discord; sudo open -a discord
+brew install --cask intellij-idea; sudo open -a "intellij idea"
+brew install --cask alt-tab; sudo open -a alttab
+brew install --cask cheatsheet; sudo open -a cheatsheet
+brew install --cask appcleaner; sudo open -a appcleaner
+brew install --cask lunar; sudo open -a lunar
 brew install maven
-
-brew install --cask handbrake
-sudo open -a handbrake
-
-brew install --cask eclipse-java 
-sudo open -a "eclipse java"
-
-brew install --cask spotify
-sudo open -a spotify
-
-brew install --cask the-unarchiver
-sudo open -a "the unarchiver"
-
-brew install --cask visual-studio
-sudo open -a "visual studio"
-
+brew install --cask handbrake; sudo open -a handbrake
+brew install --cask eclipse-java; sudo open -a "eclipse java"
+brew install --cask spotify; sudo open -a spotify
+brew install --cask the-unarchiver; sudo open -a "the unarchiver"
+brew install --cask visual-studio; sudo open -a "visual studio"
 brew install --cask oracle-jdk-javadoc
 
 brew upgrade # just to be sure :)
 
-# hopefully these app'll be installed, so i can open'em to setup some preferences
+echo "\033[0;34m Hopefully the following apps are installed at this time...opening them to setup some preferences 🔩 \033[0m"
 open -a enki
 open -a accelerate
 open -a piper
@@ -184,48 +133,31 @@ open -a keynote
 open -a numbers
 open -a dropover
 
-# CODE TO RENAME ALL FILES TO LOWERCASE...and remove directories i don't use
+echo "\033[0;34m The next script will rename all to lowercase...and remove directories i don't use 🔡 \033[0m"
 for f in *; do mv "$f" "$f.tmp"; mv "$f.tmp" "`echo $f | tr "[:upper:]" "[:lower:]"`"; done
 cd
 rm -r movies && rm bin && rmdir applicazioni
 
 # sync desktop on onedrive command
+echo "\033[0;34m Syncing OneDrive on desktop... 🌥 \033[0m"
 sudo rm -r /Users/enrico/Desktop && ln -s -n /Users/enrico/onedrive\ -\ unige.it /Users/enrico/Desktop
 
-# adding my scripts to the local user bin directory
+echo "\033[0;34m Adding my scripts to the local user bin directory 🤓 \033[0m"
 sudo mkdir ../../usr/local/bin 
 sudo cp onedrive\ -\ unige.it/my_projects/copy-of-bin/* ../../usr/local/bin
 sudo chmod -R 777 ../../usr/local/bin/my-git-push
 
-# at this point i should be sure to have installed vs for mac, so i'll remove the installer app
+echo "\033[0;34m At this point Visual Studio for mac should be installed, so I'll remove the installer app 😏 \033[0m"
 sudo rm -rf ../../Applications/Install\ Visual\ Studio\ for\ Mac.app
 
-# utm's virtual machines wait from onedrive sync + move to the right utm directory
-echo -e "\033[1;31m Wait for the UTM's virtual machine download to be done! 🛑 \033[0m"
-echo "\033[0;36m (From OneDrive sync) \033[0m"
+echo -e "\033[1;31m Wait for the UTM's virtual machine download from OneDrive to be done! 🛑 \n Next I will move VMs to the right directory 📂 \033[0m"
 read -p "Press enter to continue 😬"
 sudo mv onedrive\ -\ unige.it/Windows-arm.utm /Users/enrico/Library/Containers/com.utmapp.UTM/Data/Documents
 sudo mv onedrive\ -\ unige.it/ubuntu-arm.utm /Users/enrico/Library/Containers/com.utmapp.UTM/Data/Documents
 
-# bash script to chech if it's all installed correctly
-cd /Applications
-cd
-ls >> /Users/enrico/actual-app-list.txt
-actual=$(cat /Users/enrico/actual-app-list.txt)
-expected=$(cat "/Users/enrico/onedrive - unige.it/my_projects/automated_setup/expected-app-list.txt")
-status=$"cmp --silent -- "$actual" "$expected""
-if status; then
-    echo -e "\033[1;31m Something went wrong...check which apps is not in actual-app-list.txt 🙁 \033[0m"
-    open /Desktop/my_projects/automated_setup/expected-app-list.txt
-    open /Applications/actual-app-list.txt
-    exit
-fi
-echo -e "\033[0;32m Tutto installato correttamente 😄 \033[0m"
-rm /Users/enrico/actual-app-list.txt
 
-
-# TODO preferenze dock, quali app pinnare nella dock e mettere ordine nel launchpad
 echo "\033[0;36m Now I will edit some dock's preferences 🌟 \033[0m"
+# TODO preferenze dock, quali app pinnare nella dock e mettere ordine nel launchpad
 defaults write com.apple.dock persistent-apps -array
 # ordine app dock: safari mail foto calendario promemoria note appstore imovie monitoraggioattività enki etoro terminale teams discord mamp eclipse intellij vs vsc utm iterm2 +...
 # ...+ adguardvpn handbrake vlc firefox spotify telegram whatsapp appcleaner onedrivetrash downloads
@@ -239,6 +171,22 @@ defaults write com.apple.dock persistent-apps -array-add "<dict><key>tile-data</
 killall cfprefsd
 killall Dock
 
+
+echo "\033[0;34m At the end, I'll execute a bash script to check if it's all installed correctly 🥰 \033[0m"
+cd /Applications
+cd
+ls >> /Users/enrico/actual-app-list.txt
+actual=$(cat /Users/enrico/actual-app-list.txt)
+expected=$(cat "/Users/enrico/onedrive - unige.it/my_projects/automated_setup/expected-app-list.txt")
+status=$"cmp --silent -- "$actual" "$expected""
+if status; then
+    echo -e "\033[1;31m Something went wrong...check which apps is not in actual-app-list.txt 🙁 \033[0m"
+    open /Desktop/my_projects/automated_setup/expected-app-list.txt
+    open /Applications/actual-app-list.txt
+    exit
+fi
+echo -e "\033[0;32m Nothing went wrong! 😄 \033[0m"
+rm /Users/enrico/actual-app-list.txt
 
 read -p "Press enter to restart MacOs 🔁"
 sudo shutdown -r now
@@ -263,3 +211,9 @@ sudo shutdown -r now
 # Purple       0;35     Light Purple  1;35
 # Cyan         0;36     Light Cyan    1;36
 # Light Gray   0;37     White         1;37
+
+# curl -o /tmp/hblock 'https://raw.githubusercontent.com/hectorm/hblock/v3.4.0/hblock' \
+#   && echo '4031d86cd04fd7c6cb1b7e9acb1ffdbe9a3f84f693bfb287c68e1f1fa2c14c3b  /tmp/hblock' | shasum -c \
+#   && sudo mv /tmp/hblock /usr/local/bin/hblock \
+#   && sudo chown 0:0 /usr/local/bin/hblock \
+#   && sudo chmod 755 /usr/local/bin/hblock
